Fix bugs: API calls w/ invalid arguments
This commit is a bug fix that aims to eliminate cases where
the App component attempst to call the API with invalid arguments.
- queryUrl and forecastUrl are now initialised with a valid
  default state, instead of as empty strings
- getInitialDataUsingDefaultCity no longer takes any arguments
  - it previously took one argument, `city`, but because
    getInitialDataUsingDefaultCity is only called as an error
    callback to `getCurrentPosition()`, the only argument that
    gets passed to it is a `GeolocationPositionError`
  - instead, getInitialDataUsingDefaultCity should use the
    `city` state
- handleSubmit() now prevents queryUrl and forecastUrl from
  updating if the city name is empty

This does **not** handle the case where the user enters nonsense
input, such as 'asdfasdf', into the search bar. We still need to
figure out the error handling for that.

This commit also does the following:
- adds local weather icons, so we are no longer pulling weather
  icons from OpenWeatherMap, and
- prevents the app from trying to load a weather icon if it cannot
  retrieve an icon code.